Move metrics config to MetricsConfiguration
